debug script

+++ xyz.py
@@ -16,40 +16,48 @@ idx = 0
 bool1 = False
 bool2 = False
 bool3 = False
+bool4 = False
+bool5 = False
 
 while continue_playing:
-        with open(file_path, "r") as f:
-            lines = f.readlines()
-            last_line = lines[-1].strip()
+    with open(file_path, "r") as f:
+        lines = f.readlines()
+        last_line = lines[-1].strip()
 
-        # Check if the last line is a valid command
-        if last_line == "1":
-            bool2 = False
-            bool3 = False
-            if bool1 == False:
-                # Stop the current song, if there is one
-                if current_song is not None:
-                    os.system("killall aplay")
-                    time.sleep(1)  # wait for the current song to stop
-                
-                # Play the new song
-                current_song = songs[idx]
-                os.system(f"aplay -q {current_song} &")  # add "&" to run in the background
-                bool1 = True
+    # Check if the last line is a valid command
+    if last_line == "1":
+        bool2 = False
+        bool3 = False
+        bool4 = False
+        bool5 = False
+        if bool1 == False:
+            # Stop the current song, if there is one
+            if current_song is not None:
+                os.system("killall aplay")
+                time.sleep(1)  # wait for the current song to stop
+            
+            # Play the new song
+            current_song = songs[idx]
+            os.system(f"aplay -q {current_song} &")  # add "&" to run in the background
+            bool1 = True
 
-        elif last_line == "2":
-            bool1 = False
-            bool3 = False
-            if bool2 == False:
-                # Stop the current song, if there is one
-                if current_song is not None:                
-                    os.kill(os.getpid(), 2)  # add "&" to run in the background
-                bool2 = True
+    elif last_line == "2":
+        bool1 = False
+        bool3 = False
+        bool4 = False
+        bool5 = False
+        if bool2 == False:
+            # Stop the current song, if there is one
+            if current_song is not None:                
+                os.kill(os.getpid(), 2)  # add "&" to run in the background
+            bool2 = True
 
-        elif last_line == "3":
-            bool1 = False
-            bool2 = False
-            bool3 = False
+    elif last_line == "3":
+        bool1 = False
+        bool2 = False
+        bool4 = False
+        bool5 = False
+        if bool3 == False:
             # Stop the current song, if there is one
             if current_song is not None:
                 os.system("killall aplay")
@@ -61,11 +69,14 @@ while continue_playing:
                 idx = 3
             current_song = songs[idx]
             os.system(f"aplay -q {current_song} &")  # add "&" to run in the background
+            bool3 = True
 
-        elif last_line == "4":
-            bool1 = False
-            bool2 = False
-            bool3 = False
+    elif last_line == "4":
+        bool1 = False
+        bool2 = False
+        bool3 = False
+        bool5 = False
+        if bool4 == False:
             # Stop the current song, if there is one
             if current_song is not None:
                 os.system("killall aplay")
@@ -77,17 +88,20 @@ while continue_playing:
                 idx = 0
             current_song = songs[idx]
             os.system(f"aplay -q {current_song} &")  # add "&" to run in the background
+            bool4 = True
 
-        elif last_line == "5":
-            bool1 = False
-            bool2 = False
-            if bool3 == False:
-                # Stop the current song, if there is one
-                if current_song is not None:
-                    os.system("killall aplay")
-                    current_song = None  # reset the current song variable
-                    idx = 0
-                    bool3 = True
+    elif last_line == "5":
+        bool1 = False
+        bool2 = False
+        bool3 = False
+        bool4 = False
+        if bool5 == False:
+            # Stop the current song, if there is one
+            if current_song is not None:
+                os.system("killall aplay")
+                current_song = None  # reset the current song variable
+                idx = 0
+                bool5 = True
 
     # Wait for a few seconds before checking the file again
     time.sleep(2)
